完善预付金额界面，把我的钱包可用金额类型修改为Double类型

+++ app/src/main/java/com/xiaolian/amigo/data/network/model/dto/response/PersonalWalletDTO.java
@@ -9,6 +9,6 @@ import lombok.Data;
  */
 @Data
 public class PersonalWalletDTO {
-    private Long balance;
-    private Long prepay;
+    private Double balance;
+    private Double prepay;
 }
